// src/core/types.d.ts
export type ValueType = 'boolean' | 'integer' | 'float' | 'string' | 'datetime' | 'date' | 'time';
export type DataType = 'varchar' | 'text' | 'json' | 'int' | 'tinyint' | 'smallint' | 'double' | 'float' | 'decimal' | 'datetime' | 'date' | 'time';
export type GridColSizeType = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 'auto';
export type CharsetType = 'utf8';
// ...
export interface ButtonComponentType extends BaseComponentType {
    type: 'button';
    icon?: string | IconControlType;
    label?: string | LabelControlType;
    button?: ButtonControlType;
}
export interface CheckboxComponentType extends BaseComponentType {
    type: 'checkbox';
    icon?: string | IconControlType;
    label?: string | LabelControlType;
    checkbox?: CheckboxControlType;
}
export interface DatetimeComponentType extends BaseComponentType {
    type: 'datetime';
    icon?: string | IconControlType;
    label?: string | LabelControlType;
    datetime?: DatetimeControlType;
}
export interface InputComponentType extends BaseComponentType {
    type: 'input';
    icon?: string | IconControlType;
    label?: string | LabelControlType;
    input?: InputControlType;
}
export interface ToggleComponentType extends BaseComponentType {
    type: 'toggle';
    icon?: string | IconControlType;
    label?: string | LabelControlType;
    toggle?: ToggleControlType;
}
export type FieldComponentType = ButtonComponentType | CheckboxComponentType | DatetimeComponentType | InputComponentType | ToggleComponentType;
export interface FieldColumnType {
    type: DataType;
    length?: number;
    decimal?: number;
    /** Field accepts nulls. @default true */
    nullable?: boolean;
    autoIncrement?: boolean;
    /** Indicates whether this field has unique values, such as keys. @default false */
    unique?: boolean;
    /** Indexed field. The name follows the pattern: `IDX_entityName_fieldName` */
    index?: boolean;
    virtual?: boolean;
    comment?: string;
    charset?: CharsetType;
    collation?: string;
    binnary?: boolean;
    /** Indicates whether values are hashed/encrypted. */
    hash?: boolean;
    /** Default value. If not provided, null is used. */
    default?: any;
}
export interface FieldType {
    /** Field name. Used with `formControlName` to bind the control in reactive forms. */
    name?: string;
    /** Declares the value type of the component for the front-end.
     *
     * ```typescript
     * value: 'boolean'
     * ```
     * To set a default value:
     * ```typescript
     * value: { type: 'boolean', default: true }
     * ```
     *
     * ```typescript
     * const defaultValue = field.value.default;
     * const control = new FormControl(defaultValue, validators)
     * ```
     */
    value?: ValueType | {
        type: ValueType;
        default?: any;
    } | undefined;
    column?: FieldColumnType;
    component?: FieldComponentType;
}